removed direct link to ydb package if ydb not selected
